@@ -36,28 +36,28 @@
     "babel": "^6.3.26",
     "babel-cli": "^6.4.5",
     "babel-core": "^6.4.0",
-    "babel-eslint": "^4.1.3",
+    "babel-eslint": "^5.0.0",
     "babel-loader": "^6.2.1",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-0": "^6.3.13",
-    "basscss": "^7.0.4",
+    "basscss": "^7.1.0",
     "chai": "^3.4.0",
     "cross-env": "^1.0.7",
     "css-loader": "^0.23.0",
     "cssnext-loader": "^1.0.1",
     "es5-shim": "^4.5.6",
     "es6-promise": "^3.1.2",
-    "es6-shim": "^0.34.4",
-    "eslint": "^1.8.0",
+    "es6-shim": "^0.35.0",
+    "eslint": "^2.2.0",
     "eslint-loader": "^1.1.1",
-    "eslint-plugin-react": "^3.6.3",
+    "eslint-plugin-react": "^4.1.0",
     "file-loader": "^0.8.4",
-    "history": "^1.17.0",
-    "html-webpack-plugin": "^1.7.0",
+    "history": "^2.0.1",
+    "html-webpack-plugin": "^2.9.0",
     "immutable": "^3.7.5",
     "istanbul": "^1.0.0-alpha",
-    "jsdom": "^7.0.2",
+    "jsdom": "^8.1.0",
     "json-loader": "^0.5.3",
     "mocha": "^2.3.3",
     "postcss-loader": "^0.8.0",
