@@ -11,5 +11,5 @@ import requireHacker from 'require-hacker';
   'css',
   'svg',
 ].forEach((type) => {
-  requireHacker.hook(type, () => `module.exports = ""`);
+  requireHacker.hook(type, () => 'module.exports = ""');
 });
